resolved issue with skeleton after recent change to module
