+++ app/graphql/resolvers/person.js
@@ -2,6 +2,6 @@ import { personHandler } from '../../core/person/person.js'
 
 export default {
   Query: {
-    person: personHandler
+    Person: personHandler
   }
 }
