+++ app/index.js
@@ -1,8 +1,11 @@
 import hapiApollo from '@as-integrations/hapi'
 
-import { apolloServer, server } from './server.js'
+import { server } from './server.js'
+import { NewApolloServer } from './graphql/server.js'
 
 const init = async () => {
+  const apolloServer = await NewApolloServer()
+
   await apolloServer.start()
   await server.register({
     plugin: hapiApollo.default,
